# src/main/resources/openapi/teenymoney-api.yaml
openapi: 3.0.3
info:
  title: TeenyMoney API
  description: 부모와 자녀가 함께 사용하는 가족 금융 서비스 API
  version: 0.4.0
servers:
  - url: /
    description: 현재 Swagger가 실행 중인 서버
  - url: http://localhost:8080
    description: 로컬 개발 서버
  - url: https://www.teenymoney.kro.kr
    description: 운영 서버
...
paths:
...
  /api/v1/wallet/me:
    get:
      tags: [Wallet]
      summary: 내 지갑 잔액과 최근 거래내역 조회
      description: Access Token으로 인증된 회원의 MEMBER 타입 지갑 잔액과,
        해당 지갑의 최근 거래내역 3건을 함께 반환합니다.
      operationId: getMyWalletDetail
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletDetailApiResponse"

        "401":
          description: 토큰이 없거나 유효하지 않음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorApiResponse"


        "404":
          description: 지갑을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorApiResponse"
              example:
                success: false
                code: WALLET_NOT_FOUND
                message: 지갑을 찾을 수 없습니다.
                data: null
  /api/v1/wallet/me/transactions:
    get:
      tags: [Wallet]
      summary: 내 거래내역 목록 조회
      description: |
        Access Token으로 인증된 회원의 MEMBER 타입 지갑 거래내역을 조회합니다.
        종류(type)/기간(period)/정렬(sort) 세 조건을 조합해서 필터링하며,
        세 파라미터를 아무것도 안 보내면 기본값(전체/1개월/최신순)이 적용됩니다.
      operationId: getMyTransactions
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: type
          required: false
          schema:
            type: string
            enum: [ALL, CREDIT, DEBIT]
            default: ALL
          example: ALL
        - in: query
          name: period
          required: false
          schema:
            type: string
            enum: [WEEK, MONTH, THREE_MONTHS, SIX_MONTHS]
            default: MONTH
          example: MONTH
        - in: query
          name: sort
          required: false
          schema:
            type: string
            enum: [DESC, ASC]
            default: DESC
          example: DESC
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletTransactionListApiResponse"

        "401":
          description: 토큰이 없거나 유효하지 않음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorApiResponse"

        "404":
          description: 지갑을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorApiResponse"
              example:
                success: false
                code: WALLET_NOT_FOUND
                message: 지갑을 찾을 수 없습니다.
                data: null
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    WalletTransaction:
      type: object
      required: [id, direction, amount, balanceAfter, description, createdAt]
      properties:
        id:
          type: integer
          format: int64
          example: 1
        direction:
          type: string
          enum: [CREDIT, DEBIT]
          example: DEBIT
        amount:
          type: integer
          format: int64
          example: 3500
        balanceAfter:
          type: integer
          format: int64
          example: 146500
        description:
          type: string
          nullable: true
          example: 편의점 결제
        createdAt:
          type: string
          example: "2026-07-31 16:15:43"
    WalletDetailResponse:
      type: object
      required: [ walletId, balance, updatedAt, recentTransactions ]
      properties:
        walletId:
          type: integer
          format: int64
          example: 5
        balance:
          type: integer
          format: int64
          example: 150000
        updatedAt:
          type: string
          nullable: true
          example: "2026-07-31 16:15:43"
        recentTransactions:
          type: array
          items:
            $ref: "#/components/schemas/WalletTransaction"
    WalletDetailApiResponse:
      type: object
      required: [ success, code, message, data ]
      properties:
        success:
          type: boolean
          example: true
        code:
          type: string
          example: OK
        message:
          type: string
          example: 성공
        data:
          $ref: "#/components/schemas/WalletDetailResponse"
    WalletTransactionListApiResponse:
      type: object
      required: [ success, code, message, data ]
      properties:
        success:
          type: boolean
          example: true
        code:
          type: string
          example: OK
        message:
          type: string
          example: 성공
        data:
          type: array
          items:
            $ref: "#/components/schemas/WalletTransaction"
...
    ErrorApiResponse:
      type: object
      required: [success, code, message]
      properties:
        success:
          type: boolean
          example: false
        code:
          type: string
          example: AUTH_VERIFICATION_CODE_INVALID
        message:
          type: string
          example: 인증번호가 올바르지 않습니다.
        data:
          nullable: true
          example: null
...